+++ app/src/main/java/ru/kitfactory/mixingautopaint/domain/models/CompleteMixData.kt
@@ -1,8 +1,8 @@
 package ru.kitfactory.mixingautopaint.domain.models
 
 data class CompleteMixData(val title : String,
-                           val massPaint: Double,
-                           val massHardener: Double,
-                           val paintPlusHardener : Double,
-                           val massDiluent: Double,
-                           val massMix: Double)
+                           val massPaint: Float,
+                           val massHardener: Float,
+                           val paintPlusHardener : Float,
+                           val massDiluent: Float,
+                           val massMix: Float)
